requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.4
django-jazzmin==3.0.1
django-querycount==0.8.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
packaging==25.0
PyJWT==2.9.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
